ota_updater: try to reuse the previous interrupted OTA's base file_table if possible

+++ src/otaclient/ota_core/_updater.py
@@ -276,17 +276,24 @@ class OTAUpdater(OTAUpdateOperator):
                 base_meta_dir_on_standby_slot = (
                     self._ota_tmp_meta_on_standby / BASE_METADATA_FOLDER
                 )
-                shutil.rmtree(base_meta_dir_on_standby_slot, ignore_errors=True)
 
                 # NOTE: the file_table file in /opt/ota/image-meta MUST be prepared by otaclient,
                 #       it is not included in the OTA image, thus also not in file_table.
                 verified_base_db = None
-                if self._image_meta_dir_on_standby.is_dir():
-                    shutil.move(
-                        self._image_meta_dir_on_standby,
-                        base_meta_dir_on_standby_slot,
-                    )
+                # NOTE: if the previous OTA is interrupted, and it is base file_table assisted,
+                #       try to keep using that base file_table.
+                if base_meta_dir_on_standby_slot.is_dir():
                     verified_base_db = find_saved_fstable(base_meta_dir_on_standby_slot)
+                else:
+                    shutil.rmtree(base_meta_dir_on_standby_slot, ignore_errors=True)
+                    if self._image_meta_dir_on_standby.is_dir():
+                        shutil.move(
+                            self._image_meta_dir_on_standby,
+                            base_meta_dir_on_standby_slot,
+                        )
+                        verified_base_db = find_saved_fstable(
+                            base_meta_dir_on_standby_slot
+                        )
 
                 _inplace_mode_params = DeltaGenParams(
                     file_table_db_helper=_fst_db_helper,
